Extract e3 search code
