+++ turbo/jsonrpc/otterscan_api.go
@@ -5,10 +5,8 @@ import (
 	"errors"
 	"fmt"
 	"math/big"
-	"slices"
 
 	hexutil2 "github.com/ledgerwatch/erigon-lib/common/hexutil"
-	"github.com/ledgerwatch/erigon/cmd/state/exec3"
 	"golang.org/x/sync/errgroup"
 
 	"github.com/holiman/uint256"
@@ -16,9 +14,6 @@ import (
 	"github.com/ledgerwatch/erigon-lib/common"
 	"github.com/ledgerwatch/erigon-lib/common/hexutility"
 	"github.com/ledgerwatch/erigon-lib/kv"
-	"github.com/ledgerwatch/erigon-lib/kv/iter"
-	"github.com/ledgerwatch/erigon-lib/kv/order"
-	"github.com/ledgerwatch/erigon-lib/kv/rawdbv3"
 	"github.com/ledgerwatch/erigon/consensus"
 	"github.com/ledgerwatch/erigon/core"
 	"github.com/ledgerwatch/erigon/core/rawdb"
@@ -29,7 +24,6 @@ import (
 	"github.com/ledgerwatch/erigon/turbo/adapter/ethapi"
 	"github.com/ledgerwatch/erigon/turbo/rpchelper"
 	"github.com/ledgerwatch/erigon/turbo/transactions"
-	"github.com/ledgerwatch/log/v3"
 )
 
 // API_LEVEL Must be incremented every time new additions are made
@@ -259,48 +253,6 @@ func (api *OtterscanAPIImpl) SearchTransactionsBefore(ctx context.Context, addr 
 	return &TransactionsWithReceipts{txs, receipts, isFirstPage, !hasMore}, nil
 }
 
-func createBackwardTxNumIter(tx kv.TemporalTx, addr common.Address, fromTxNum int) (*rawdbv3.MapTxNum2BlockNumIter, error) {
-	// unbounded limit on purpose, since there could be e.g. block rewards system txs, we limit
-	// results later
-	itTo, err := tx.IndexRange(kv.TracesToIdx, addr[:], fromTxNum, -1, order.Desc, kv.Unlim)
-	if err != nil {
-		return nil, err
-	}
-	itFrom, err := tx.IndexRange(kv.TracesFromIdx, addr[:], fromTxNum, -1, order.Desc, kv.Unlim)
-	if err != nil {
-		return nil, err
-	}
-	txNums := iter.Union[uint64](itFrom, itTo, order.Desc, kv.Unlim)
-	return rawdbv3.TxNums2BlockNums(tx, txNums, order.Desc), nil
-}
-
-func (api *OtterscanAPIImpl) searchTransactionsBeforeV3(tx kv.TemporalTx, ctx context.Context, addr common.Address, fromBlockNum uint64, pageSize uint16) (*TransactionsWithReceipts, error) {
-	isFirstPage := false
-	if fromBlockNum == 0 {
-		isFirstPage = true
-	} else {
-		// Internal search code considers blockNum [including], so adjust the value
-		fromBlockNum--
-	}
-	fromTxNum := -1
-	if fromBlockNum != 0 {
-		// from == 0 == magic number which means last; reproduce bug-compatibility for == 1
-		// with e2 for now
-		_txNum, err := rawdbv3.TxNums.Max(tx, fromBlockNum)
-		if err != nil {
-			return nil, err
-		}
-		fromTxNum = int(_txNum)
-	}
-
-	txs, receipts, hasMore, err := api.buildSearchResults(ctx, tx, createBackwardTxNumIter, addr, fromTxNum, pageSize)
-	if err != nil {
-		return nil, err
-	}
-
-	return &TransactionsWithReceipts{txs, receipts, isFirstPage, !hasMore}, nil
-}
-
 // Search transactions that touch a certain address.
 //
 // It searches forward a certain block (excluding); the results are sorted descending.
@@ -392,134 +344,6 @@ func (api *OtterscanAPIImpl) SearchTransactionsAfter(ctx context.Context, addr c
 		receipts[i], receipts[lentxs-1-i] = receipts[lentxs-1-i], receipts[i]
 	}
 	return &TransactionsWithReceipts{txs, receipts, !hasMore, isLastPage}, nil
-}
-
-func createForwardTxNumIter(tx kv.TemporalTx, addr common.Address, fromTxNum int) (*rawdbv3.MapTxNum2BlockNumIter, error) {
-	// unbounded limit on purpose, since there could be e.g. block rewards system txs, we limit
-	// results later
-	itTo, err := tx.IndexRange(kv.TracesToIdx, addr[:], fromTxNum, -1, order.Asc, kv.Unlim)
-	if err != nil {
-		return nil, err
-	}
-	itFrom, err := tx.IndexRange(kv.TracesFromIdx, addr[:], fromTxNum, -1, order.Asc, kv.Unlim)
-	if err != nil {
-		return nil, err
-	}
-	txNums := iter.Union[uint64](itFrom, itTo, order.Asc, kv.Unlim)
-	return rawdbv3.TxNums2BlockNums(tx, txNums, order.Asc), nil
-
-}
-
-func (api *OtterscanAPIImpl) searchTransactionsAfterV3(tx kv.TemporalTx, ctx context.Context, addr common.Address, fromBlockNum uint64, pageSize uint16) (*TransactionsWithReceipts, error) {
-	isLastPage := false
-	fromTxNum := -1
-	if fromBlockNum == 0 {
-		isLastPage = true
-	} else {
-		// Internal search code considers blockNum [including], so adjust the value
-		_txNum, err := rawdbv3.TxNums.Min(tx, fromBlockNum+1)
-		if err != nil {
-			return nil, err
-		}
-		fromTxNum = int(_txNum)
-	}
-
-	txs, receipts, hasMore, err := api.buildSearchResults(ctx, tx, createForwardTxNumIter, addr, fromTxNum, pageSize)
-	if err != nil {
-		return nil, err
-	}
-	slices.Reverse(txs)
-	slices.Reverse(receipts)
-
-	return &TransactionsWithReceipts{txs, receipts, !hasMore, isLastPage}, nil
-}
-
-type txNumsIterFactory func(tx kv.TemporalTx, addr common.Address, fromTxNum int) (*rawdbv3.MapTxNum2BlockNumIter, error)
-
-func (api *OtterscanAPIImpl) buildSearchResults(ctx context.Context, tx kv.TemporalTx, iterFactory txNumsIterFactory, addr common.Address, fromTxNum int, pageSize uint16) ([]*RPCTransaction, []map[string]interface{}, bool, error) {
-	chainConfig, err := api.chainConfig(ctx, tx)
-	if err != nil {
-		return nil, nil, false, err
-	}
-
-	txNumsIter, err := iterFactory(tx, addr, fromTxNum)
-	if err != nil {
-		return nil, nil, false, err
-	}
-
-	exec := exec3.NewTraceWorker(tx, chainConfig, api.engine(), api._blockReader, nil)
-	var blockHash common.Hash
-	var header *types.Header
-	txs := make([]*RPCTransaction, 0, pageSize)
-	receipts := make([]map[string]interface{}, 0, pageSize)
-	resultCount := uint16(0)
-
-	mustReadHeader := true
-	for txNumsIter.HasNext() {
-		txNum, blockNum, txIndex, isFinalTxn, blockNumChanged, err := txNumsIter.Next()
-		if err != nil {
-			return nil, nil, false, err
-		}
-		mustReadHeader = mustReadHeader || blockNumChanged
-		if isFinalTxn {
-			continue
-		}
-
-		if mustReadHeader {
-			if header, err = api._blockReader.HeaderByNumber(ctx, tx, blockNum); err != nil {
-				return nil, nil, false, err
-			}
-			if header == nil {
-				log.Warn("[rpc] header is nil", "blockNum", blockNum)
-				continue
-			}
-			blockHash = header.Hash()
-			exec.ChangeBlock(header)
-			mustReadHeader = false
-		}
-
-		txn, err := api._txnReader.TxnByIdxInBlock(ctx, tx, blockNum, txIndex)
-		if err != nil {
-			return nil, nil, false, err
-		}
-		if txn == nil {
-			log.Warn("[rpc] txn not found", "blockNum", blockNum, "txIndex", txIndex)
-			continue
-		}
-		res, err := exec.ExecTxn(txNum, txIndex, txn)
-		if err != nil {
-			return nil, nil, false, err
-		}
-		rawLogs := exec.GetLogs(txIndex, txn)
-		rpcTx := NewRPCTransaction(txn, blockHash, blockNum, uint64(txIndex), header.BaseFee)
-		txs = append(txs, rpcTx)
-		receipt := &types.Receipt{
-			Type:              txn.Type(),
-			GasUsed:           res.UsedGas,
-			CumulativeGasUsed: res.UsedGas, // TODO: cumulative gas is wrong, wait for cumulative gas index fix
-			TransactionIndex:  uint(txIndex),
-			BlockNumber:       header.Number,
-			BlockHash:         blockHash,
-			Logs:              rawLogs,
-		}
-		if res.Failed() {
-			receipt.Status = types.ReceiptStatusFailed
-		} else {
-			receipt.Status = types.ReceiptStatusSuccessful
-		}
-
-		mReceipt := marshalReceipt(receipt, txn, chainConfig, header, txn.Hash(), true)
-		mReceipt["timestamp"] = header.Time
-		receipts = append(receipts, mReceipt)
-
-		resultCount++
-		if resultCount >= pageSize {
-			break
-		}
-	}
-
-	hasMore := txNumsIter.HasNext()
-	return txs, receipts, hasMore, nil
 }
 
 func (api *OtterscanAPIImpl) traceBlocks(ctx context.Context, addr common.Address, chainConfig *chain.Config, pageSize, resultCount uint16, callFromToProvider BlockProvider) ([]*TransactionsWithReceipts, bool, error) {
